MAINT: Create CustomTheme file
