+++ src/index.tsx
@@ -1,56 +1,22 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import CssBaseline from '@material-ui/core/CssBaseline'
-import { createTheme, MuiThemeProvider } from '@material-ui/core/styles'
 
 import './index.css'
 import App from './components/App'
+import CustomTheme from './theme/CustomTheme'
 import reportWebVitals from './reportWebVitals'
 
 const Index: React.FC = () => {
   const [isLightMode, setIsLightMode] = React.useState<boolean>(true)
   const handleThemeMode = () => setIsLightMode(!isLightMode)
 
-  // TODO: Move to other file
-  const customTheme = createTheme({
-    mixins: {
-      toolbar: {
-        minHeight: 48,
-      },
-    },
-    palette: {
-      type: isLightMode ? 'light' : 'dark',
-    },
-    props: {
-      MuiTextField: {
-        variant: 'outlined',
-      },
-      MuiList: {
-        dense: true,
-      },
-      MuiCheckbox: {
-        color: 'primary',
-      },
-      MuiRadio: {
-        color: 'primary',
-      },
-      MuiSwitch: {
-        color: 'primary',
-      },
-    },
-    typography: {
-      button: {
-        textTransform: 'none',
-      },
-    },
-  })
-
   return (
     <React.StrictMode>
-      <MuiThemeProvider theme={customTheme}>
+      <CustomTheme isLightMode={isLightMode}>
         <CssBaseline />
         <App handleTheme={handleThemeMode} />
-      </MuiThemeProvider>
+      </CustomTheme>
     </React.StrictMode>
   )
 }
